requirements.txt
streamlit
playwright
pandas
python-dotenv
xlsxwriter
